Move the schema-level table lookup out of storage_lookup.h
storage_lookup.h sat at the top level and mixed two things: generic facilities
for the tuple of database objects, and the algorithms that search that tuple
for a table. Split them and put each where it belongs.

The lookup algorithms move to schema/algorithms/table_lookup.h, alongside
sync_order.h. They search across the collection of database objects rather
than classifying a node already in hand, which is what the vocabulary layer
does. Rename storage_pick_table[_t] and storage_find_table[_t] accordingly to
schema_pick_table[_t] and schema_find_table[_t]: these operate on the schema,
the mapped database objects, not on the storage_t object.

The generic half (db_objects_tuple, is_db_objects, db_objects_for_expression)
is not an algorithm and moves to schema/db_objects.h. Include sites now pull
in whichever of the two headers they actually use.

Drop is_storage, and with it the forward declaration of storage_t that existed
only to serve it. It had no users other than the static test asserting on it.

Record at file level that lookup covers tables in the wide SQL sense - base
tables, views and virtual tables - and deliberately not indexes or triggers,
whose void object_type object_type_matches filters out.

Also drop an unused include from udf_existence_checker.h.
