+++ dev/column_names_getter.h
@@ -15,7 +15,7 @@
 #include "error_code.h"
 #include "mapped_type_proxy.h"
 #include "alias_traits.h"
-#include "storage_lookup.h"  //  pick_table
+#include "schema/algorithms/table_lookup.h"  //  pick_table
 #include "util.h"  // quote_identifier
 #include "vocabulary/node_traits.h"
 #include "vocabulary/node_algorithms.h"  // access_column_expression
