+++ dev/column_result.h
@@ -24,6 +24,7 @@
 #include "alias.h"
 #include "cte_types.h"
 #include "storage_traits.h"
+#include "schema/algorithms/table_lookup.h"  // schema_pick_table_t
 #include "function.h"
 #include "ast/special_keywords.h"
 #include "ast/cast.h"
@@ -326,7 +327,7 @@ namespace sqlite_orm::internal {
 #if (SQLITE_VERSION_NUMBER >= 3008003) && defined(SQLITE_ORM_WITH_CTE)
     template<class DBOs, class Moniker, class ColAlias>
     struct column_result_t<DBOs, column_pointer<Moniker, alias_holder<ColAlias>>, void> {
-        using table_type = storage_pick_table_t<Moniker, DBOs>;
+        using table_type = schema_pick_table_t<Moniker, DBOs>;
         using cte_mapper_type = cte_mapper_type_t<table_type>;
 
         // lookup ColAlias in the final column references
