+++ dev/object_from_column_builder.h
@@ -12,7 +12,7 @@
 #include "vocabulary/node_fwd.h"  // column_field
 #include "table_reference.h"
 #include "row_extractor.h"
-#include "storage_lookup.h"
+#include "schema/algorithms/table_lookup.h"
 
 namespace sqlite_orm::internal {
     struct object_from_column_builder_base {
