+++ dev/schema/algorithms/table_lookup.h
@@ -1,43 +1,32 @@
 #pragma once
 
+/** @file Lookup of a table definition within a tuple of database objects.
+ *
+ *  These are schema-level algorithms: they search *across* the collection of database objects
+ *  to locate the one mapping a given lookup type, as opposed to classifying a node already in hand,
+ *  which is what the vocabulary layer does. They consume the vocabulary layer to do so.
+ *
+ *  "Table" is meant in the wide SQL table sense here, covering base tables, views
+ *  and virtual tables alike. Indexes and triggers are deliberately not covered:
+ *  their `object_type` is void, which `object_type_matches` filters out, so a lookup
+ *  answers whether a type is mapped as a table - not whether it occurs in the schema at all.
+ */
+
 #ifndef SQLITE_ORM_IMPORT_STD_MODULE
-#include <type_traits>  //  std::true_type, std::false_type, std::remove_const, std::enable_if, std::is_base_of, std::is_void
+#include <type_traits>  //  std::true_type, std::false_type, std::remove_const, std::enable_if, std::is_same, std::is_void
 #include <tuple>  // std::tuple_size, std::get
 #include <utility>  //  std::index_sequence, std::make_index_sequence
 #endif
 
-#include "functional/cxx_type_traits_polyfill.h"
-#include "vocabulary/node_projections.h"
-#include "type_traits.h"
+#include "../../functional/cxx_type_traits_polyfill.h"
+#include "../../vocabulary/node_projections.h"
+#include "../../type_traits.h"
+#include "../db_objects.h"
 
 namespace sqlite_orm::internal {
-    template<class... DBO>
-    struct storage_t;
-
-    template<class... DBO>
-    using db_objects_tuple = std::tuple<DBO...>;
-
-    template<class T>
-    struct is_storage : std::false_type {};
-
-    template<class... DBO>
-    struct is_storage<storage_t<DBO...>> : std::true_type {};
-    template<class... DBO>
-    struct is_storage<const storage_t<DBO...>> : std::true_type {};
-
-    template<class T>
-    struct is_db_objects : std::false_type {};
-
-    template<class... DBO>
-    struct is_db_objects<std::tuple<DBO...>> : std::true_type {};
-    // note: cannot use `db_objects_tuple` alias template because older compilers have problems
-    // to match `const db_objects_tuple`.
-    template<class... DBO>
-    struct is_db_objects<const std::tuple<DBO...>> : std::true_type {};
-
     /**
      *  `std::true_type` if given object is mapped, `std::false_type` otherwise.
-     *  
+     *
      *  Note: unlike base_table<>, index_t<>::object_type and trigger_t<>::object_type is always void.
      */
     template<typename DBO, typename Lookup>
@@ -62,56 +51,56 @@ namespace sqlite_orm::internal {
 
     /**
      *  SFINAE friendly facility to pick a table definition (`base_table`) from a tuple of database objects.
-     *  
+     *
      *  Lookup - mapped data type
      *  Seq - index sequence matching the number of DBOs
      *  DBOs - db_objects_tuple type
      */
     template<class Lookup, class Seq, class DBOs>
-    struct storage_pick_table;
+    struct schema_pick_table;
 
     template<class Lookup, size_t... Ix, class... DBO>
-    struct storage_pick_table<Lookup, std::index_sequence<Ix...>, db_objects_tuple<DBO...>>
+    struct schema_pick_table<Lookup, std::index_sequence<Ix...>, db_objects_tuple<DBO...>>
         : enable_found_table<Lookup, Ix, DBO>... {};
 
     /**
      *  SFINAE friendly facility to pick a table definition (`base_table`) from a tuple of database objects.
-     *  
+     *
      *  Lookup - 'table' type, mapped data type
      *  DBOs - db_objects_tuple type, possibly const-qualified
      */
     template<class Lookup, class DBOs>
-    using storage_pick_table_t = typename storage_pick_table<Lookup,
-                                                             std::make_index_sequence<std::tuple_size<DBOs>::value>,
-                                                             std::remove_const_t<DBOs>>::type;
+    using schema_pick_table_t = typename schema_pick_table<Lookup,
+                                                           std::make_index_sequence<std::tuple_size<DBOs>::value>,
+                                                           std::remove_const_t<DBOs>>::type;
 
     /**
      *  Find a table definition (`base_table`) from a tuple of database objects;
      *  `std::nonesuch` if not found.
-     *  
+     *
      *  DBOs - db_objects_tuple type
      *  Lookup - mapped data type
      */
     template<class Lookup, class DBOs>
-    struct storage_find_table : polyfill::detected<storage_pick_table_t, Lookup, DBOs> {};
+    struct schema_find_table : polyfill::detected<schema_pick_table_t, Lookup, DBOs> {};
 
     /**
      *  Find a table definition (`base_table`) from a tuple of database objects;
      *  `std::nonesuch` if not found.
-     *  
+     *
      *  DBOs - db_objects_tuple type, possibly const-qualified
      *  Lookup - mapped data type
      */
     template<class Lookup, class DBOs>
-    using storage_find_table_t = typename storage_find_table<Lookup, std::remove_const_t<DBOs>>::type;
+    using schema_find_table_t = typename schema_find_table<Lookup, std::remove_const_t<DBOs>>::type;
 
 #ifndef SQLITE_ORM_BROKEN_VARIADIC_PACK_EXPANSION
     template<class DBOs, class Lookup, class SFINAE = void>
     struct is_mapped : std::false_type {};
     template<class DBOs, class Lookup>
-    struct is_mapped<DBOs, Lookup, polyfill::void_t<storage_pick_table_t<Lookup, DBOs>>> : std::true_type {};
+    struct is_mapped<DBOs, Lookup, polyfill::void_t<schema_pick_table_t<Lookup, DBOs>>> : std::true_type {};
 #else
-    template<class DBOs, class Lookup, class SFINAE = storage_find_table_t<Lookup, DBOs>>
+    template<class DBOs, class Lookup, class SFINAE = schema_find_table_t<Lookup, DBOs>>
     struct is_mapped : std::true_type {};
     template<class DBOs, class Lookup>
     struct is_mapped<DBOs, Lookup, polyfill::nonesuch> : std::false_type {};
@@ -125,21 +114,13 @@ namespace sqlite_orm::internal {
 namespace sqlite_orm::internal {
     /**
      *  Pick the table definition for the specified lookup type from the given tuple of schema objects.
-     *  
+     *
      *  Note: This function requires Lookup to be mapped, otherwise it is removed from the overload resolution set.
      */
     template<class Lookup, class DBOs, satisfies<is_mapped, DBOs, Lookup> = true>
     auto& pick_table(DBOs& dbObjects) {
-        using table_type = storage_pick_table_t<Lookup, DBOs>;
+        using table_type = schema_pick_table_t<Lookup, DBOs>;
         return std::get<table_type>(dbObjects);
-    }
-
-    /**
-     *  Return passed in DBOs.
-     */
-    template<class DBOs, class E, satisfies<is_db_objects, DBOs> = true>
-    decltype(auto) db_objects_for_expression(DBOs& dbObjects, const E&) {
-        return dbObjects;
     }
 
     template<class Lookup, class DBOs, satisfies<is_db_objects, DBOs> = true>
