+++ tests/static_tests/select_return_type.cpp
@@ -33,12 +33,6 @@ TEST_CASE("Select return types") {
         STATIC_REQUIRE(internal::is_mapped_v<decltype(storage)::db_objects_type, User>);
         STATIC_REQUIRE_FALSE(internal::is_mapped_v<decltype(storage)::db_objects_type, Visit>);
 
-        //  test is_storage
-        STATIC_REQUIRE(internal::is_storage<decltype(storage)>::value);
-        STATIC_REQUIRE_FALSE(internal::is_storage<User>::value);
-        STATIC_REQUIRE_FALSE(internal::is_storage<int>::value);
-        STATIC_REQUIRE_FALSE(internal::is_storage<void>::value);
-
         auto storage2 = make_storage(
             "",
             make_table("visits", make_column("id", &Visit::id, primary_key()), make_column("date", &Visit::date)));
